plugins_demos/outside_ocr_loaders/deps/chineseocr_lite/requirements.txt
tornado==6.4.1
numpy==1.20.0
opencv-python==4.10.0.84
onnxruntime==1.10.0
shapely==2.0.6
pyclipper==1.3.0.post6
Pillow==9.5.0
